pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "krishirakshak-backend"
version = "0.1.0"
description = "KrishiRakshak FastAPI backend for Android/Web/Voice"
authors = [{name="KrishiRakshak Team"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic-settings>=2.0.3",
  "python-multipart>=0.0.9",
  "sqlalchemy>=2.0.32",
  "sqlmodel>=0.0.21",
  "asyncpg>=0.29.0",
  "aiosqlite>=0.20.0",
  "greenlet>=3.0.0",
  "alembic>=1.13.3",
  "passlib[bcrypt]>=1.7.4",
  "pyjwt[crypto]>=2.9.0",
  "celery[redis]>=5.4.0",
  "redis>=5.0.8",
  "boto3>=1.35.0",
  "minio>=7.2.7",
  "httpx>=0.27.0",
  "aiofiles>=24.1.0",
  "tenacity>=9.0.0",
  "orjson>=3.10.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-asyncio>=0.24.0",
  "anyio>=4.4.0",
  "requests>=2.32.3",
  "ruff>=0.6.8",
]

[tool.ruff]
line-length = 100
select = ["E","F","I","UP","B","SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"
